requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.2
fonttools==4.58.0
frozenlist==1.7.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.6.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
propcache==0.3.2
pycryptodome==3.23.0
pyparsing==3.2.3
python-binance==1.0.29
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
